Create useLayerMenu
